modules: Avoid unwritable log.pipe
If a privileged module is the first to execute for a given host and it
sources module.sh, then log.pipe is created with ownership such that a
later unprivileged module may not be able to write to it.

This is not an issue with log.socket since it is created by ssh(1)
during forwarding.